Add rand as dependency. Always generate a new cache version.
Each time the app is built will result in a new service worker being
installed and thus forcing the cache to be re-fetched the next time
the app is reloaded. This is not ideal, but until there is a way to
prompt the user to initiate an update to the latest version this will
have to do.